[search-ui] allow multiple triggers, style tweaks
