+++ packages/search-ui/src/components/CommandMenuTrigger.tsx
@@ -16,22 +16,8 @@ export const CommandMenuTrigger = ({ setOpen, className }: Props) => {
     setIsMac(typeof navigator !== 'undefined' && isAppleDevice());
   }, []);
 
-  useEffect(() => {
-    if (isMac !== null) {
-      const keyDownListener = (e: KeyboardEvent) => {
-        if (e.key === 'k' && (isMac ? e.metaKey : e.ctrlKey)) {
-          e.preventDefault();
-          setOpen((open) => !open);
-        }
-      };
-      document.addEventListener('keydown', keyDownListener, false);
-      return () => document.removeEventListener('keydown', keyDownListener);
-    }
-  }, [isMac]);
-
   return (
     <Button
-      id="cmdk-trigger"
       theme="secondary"
       leftSlot={<SearchSmIcon className="icon-md text-icon-secondary" />}
       rightSlot={
@@ -42,7 +28,7 @@ export const CommandMenuTrigger = ({ setOpen, className }: Props) => {
           </div>
         ) : undefined
       }
-      className={mergeClasses('bg-default pl-2.5 pr-3 border border-default shadow-xs min-h-[40px]', className)}
+      className={mergeClasses('cmdk-trigger bg-default pl-2.5 pr-3 border border-default shadow-xs min-h-[40px]', className)}
       onClick={() => setOpen(true)}>
       <p className="text-secondary font-normal leading-normal text-xs">Search</p>
     </Button>
